Initial wiring of sample echo command message.
